Phase 4: styles.css split batch 4 — arcade catalog surface
Extracts ArcadeCatalog.tsx's CSS into two files, split at the natural
toolbar-chrome/grid-content boundary since the combined content (~650
lines) is well over the 500-line hard cap:

- surfaces/catalog/catalog-toolbar.css: header, jumpbar, pager, toolbar,
  sort menu (.arcade-*, #browse-everything/#play-anchor/#party-rail,
  .catalog-jumpbar, .jump-chip*, .catalog-pager*, .catalog-toolbar,
  .catalog-sort-*)
- surfaces/catalog/catalog.css: grid, cards, media, previews, moments
  (.catalog-grid/-card/-media/-preview*/-moments/-moment/-overlay,
  .preview-toggle, .touch-warning-pill, .genre-pill, .card-title/-copy/
  -author/-actions)

Both imported only from ArcadeCatalog.tsx.

Verified every selector individually before extracting, including a
full-file grep for the class prefix (not just the initially-found range)
per the lesson from batch 2's non-contiguous mobile-override bug. Found
and fixed the same bug class again: a stray @media (max-width: 768px)
block overriding .catalog-grid/.catalog-toolbar/.catalog-filter-trigger/
.catalog-sort-trigger lived in a large block shared by many other
surfaces, far from the main chunk. Moved those three rules into their
own file-local @media blocks (appended after the base rules in each
respective file) so they always win the cascade regardless of load order
elsewhere, with a regression test.

Also caught and correctly kept in scope: an #browse-everything/
#play-anchor/#party-rail ID-selector group that IS ArcadeCatalog-owned
(verified against JSX) — the opposite finding from batch 3's #hero-prompt
near-miss, confirming the "verify every selector individually" discipline
catches both false negatives and false positives.

Deliberately left .catalog-refresh-error*/.catalog-state in styles.css:
shared with PartyPage.tsx/SubmissionStatusView.tsx, same pattern as
batch 2's propose-*/proposal-*.

Moved styles.catalogGrid.test.ts to surfaces/catalog/catalog.css.test.ts,
now reading catalog.css/catalog-toolbar.css and shrunk to zero comment-
prose debt (was carrying 44 words of pre-existing debt in the baseline;
convention is to shrink debt in any file touched, not carry it forward).
